requirements.txt
celery==5.3.6
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-crontab==0.7.1
django-phonenumber-field==7.2.0
gunicorn==20.1.0
django-filter==23.4
django-choices==2.0.0
pre-commit==3.5.0
phonenumbers==8.13.26
requests==2.31.0
boto3==1.34.36
botocore==1.34.50
inflect==7.0.0
psycopg2-binary==2.9.9
slack_bolt==1.19.0
slack_sdk==3.30.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain-openai